"Submissions and Claims": revert Whatsapp usernames
They only work after a fix check-api exports the proper rows. In the
meantime, we want clients to see _something_.

Now, on production, there are two queries that work (Submissions and
Claims; Tasks and Metadata) and two that work pending the deployment of
https://github.com/meedan/check-api/pull/637 (Tipline Conversations,
Items)

("Items" is like "Submissions and Claims", except it doesn't try to join
conversations. Now that "Tipline Conversations" is its own query,
Workbench is a better place to join.)